LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/events-004-expected.txt


PASS repeating lists
FAIL truncating lists assert_approx_equals: expected 0.02 +/- 0.0005 but got 0.01
